[components][sandbox] Land type-specific routers in component manager (Part 3)
Delete the function to convert SpecificRouter into Router, and update
all places where this is used to use SpecificRouter directly.

Also: When creating a scoped logger for a component, use the
`program_input_dict` instead of the legacy routing API to connect to the
LogSink.

Bug: 360221168

Change-Id: I4ba474c72e714435d07ecef2bc9e7706ff0c81fe
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1108852
Reviewed-by: Claire Gonyeo <cgonyeo@google.com>
Commit-Queue: Gary Bressler <geb@google.com>