[starnix] Add temporary component features for tests
These features will be present in the SELinux tests until we either
decouple the test and container cml files, or we add logic to the test
runner to modify the program block. This change does not do the latter
because that has caused a lot of confusion in the past.

Bug: 371368307
Change-Id: I1b5acb536ab47e736f31c52ef21b1157e6a87c1b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1144872
Reviewed-by: Wez <wez@google.com>
Reviewed-by: Adam Perry <adamperry@google.com>
Commit-Queue: Kevin Lindkvist <lindkvist@google.com>
Fuchsia-Auto-Submit: Kevin Lindkvist <lindkvist@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>