[plat-config] add health check field to component configuration
The health check field represents a list of opted-in components
that can potentially block an update. It is currently part of
SWD codebase but is being redesigned to be in component framework.

The product sets an enum corresponding to a component that has
implemented the health check service and the enum is translated
to a component moniker to be used in the framework.

Bug: 369128580

Change-Id: Ie51262ae5ffd87a02c1908477ac68a84be3d52e2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1122712
Reviewed-by: David Gilhooley <dgilhooley@google.com>
Commit-Queue: Olivia Nordquist <nolivia@google.com>
API-Review: Aaron Wood <aaronwood@google.com>
Reviewed-by: Aaron Wood <aaronwood@google.com>