[test_list_tool] Fix handling file relative blobs
File relative blobs in package manifests are pre-processed by the
manifest parsing library to be relative to the working directory.
"working_dir" relative manifests are actually relative to the build
directory, which is assumed to be the working dir for package
operations. test_list_tool violates this assumption when run from fx
test because it is relative to the source directory.

While the relative path computation is technically correct, the
incorrect defaults were corrected by prepending the build directory in
loading. This change ensures that the build directory is not prepended
twice by skipping that operation if the prefix is already set.

One result was that any package using "file" relative blobs failed to
get the correct directory set when using fx --dir=<> test, which is now
fixed.

Fixed: 375256332
Change-Id: I71cbd76ca12ff4d10a5b52dd71d28345568504e0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1144841
Commit-Queue: Christopher Johnson <crjohns@google.com>
Reviewed-by: Miguel Flores <miguelfrde@google.com>
Reviewed-by: Gurjant Kalsi <gkalsi@google.com>
Fuchsia-Auto-Submit: Christopher Johnson <crjohns@google.com>