[ffx] enable port:ip targets
Instead of requiring `ffx target add <target>`, just using a port:ip
should be enough for the daemon to try to reach the specified target.
(This functionality already existed, but only worked when discovery
was disabled. This change makes it so it is always used.)

Bug: 354219263
Change-Id: I87caa0efc08f7408cddef9f6255a9f91c76a2023
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1143503
Reviewed-by: Andrew Davies <awdavies@google.com>
Commit-Queue: Steven Grady <slgrady@google.com>