[zxio][netstack] Refactor UDP prelude constant sizes
The networking stack and client library (zxio) have to agree with each
other on the size of the UDP prelude messages passed in datagram socket
buffers. This constant must be available to C++ code in zxio and to Go
code in netstack. To arrange for this, there are constants defined in C
and exposed via CGo. This rearranges the definitions of these constants
into the following layers:

1. zxio defines the sizes that it uses as C++ inline constexpr variables
   in a udp_socket_private.h header. This header is restricted by visibility
   rules to be accessible only by zxio and parts of netstack.
2. src/connectivity/network/netstack/udp_serde relies on its C
   compatible definition for the routines exposed via CGo and contains
   static assertions that these constants match the values used by zxio.

As a result zxio no longer has a dependency on //src/connectivity/.. for
these particular constants and the zxio code can use constexpr variables
with the values available inline. This makes the values available for
use in constexpr contexts and provides a (minor) opportunity for further
toolchain optimizations.

Change-Id: I2cbe82fb594be1a47e38d899b1f85ab1376dd3e4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1130269
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Reviewed-by: Adam Barth <abarth@google.com>
Fuchsia-Auto-Submit: James Robinson <jamesr@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>