[storage][conformance] Update Rights struct
The Rights struct in the storage conformance library should only be used
for Rights. Update it so that it is created with rights, and also add
functions so that it will return the Open3 right flags.

There were some conformance tests that did not run previously, probably
due to confusion of `valid_combos_with/without(flags)` which should not
be used for non-rights flags. Now that they are running, there were some
tests that fail and this CL also include fixes for them:

* stream reading past maximum buffer return OUT_OF_RANGE error
  (previously was returning INVALID_ARGS)
* the block device in the test harness for minfs does not have enough
  blocks and inodes
* truncate tests should only run if append is supported.

Fixed: 293951181
Test: fx test storage_conformance_test
Change-Id: I66c8c38748e66adcd5053a3bbaa9d59439342f50
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1143692
Commit-Queue: Rui Qi Sim <rqsim@google.com>
Reviewed-by: Brandon Castellano <bcastell@google.com>
Reviewed-by: Chris Suter <csuter@google.com>