[gpio][core][test] Don't assume that pinimpl calls will be inlined
gpio-test currently uses RunUntilIdle() in a few cases that test that
the GPIO core driver calls out to the mock pinimpl driver. The mock
driver is on the environment dispatcher, so this is only guaranteed to
work if the calls are inlined. Switch to Run() instead of RunUntilIdle()
so that the tests will work regardless of how the FIDL calls are made.

Mutliply: fuchsia-pkg://fuchsia.com/gpio-test#meta/gpio-test.cm
Change-Id: I233e02d68acf860a7ae307eab9f1e5ce3ce5476b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1143348
Commit-Queue: Braden Kell <bradenkell@google.com>
Reviewed-by: Ruby Zhuang <rdzhuang@google.com>
Reviewed-by: Andres Oportus <andresoportus@google.com>
Fuchsia-Auto-Submit: Braden Kell <bradenkell@google.com>