[fxfs] handle multiple checksums per block in journal replay
When writing to a block on disk, checksums are emitted as journal
records regardless of whether that block has been written to before.
This means overwrite mode blocks that are written to multiple times
between journal flushes will have multiple checksums emitted for them.

In this case, journal replay should succeed if any of the checksums
match the data. This is so if later writes go through on that block, it
won't cause replays of earlier transactions to fail. While that would
technically be consistent, it would potentially lose a lot of data
because we would end replay earlier than we should have.

Bug: 293943124
Test: fx test fxfs-tests fxfs-platform-test
Change-Id: I0b5e19a59a9e6abee3518759e0170491cf9168b3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1135552
Fuchsia-Auto-Submit: Stephen Demos <sdemos@google.com>
Reviewed-by: James Sullivan <jfsulliv@google.com>
Reviewed-by: Chris Suter <csuter@google.com>
Commit-Queue: Stephen Demos <sdemos@google.com>