[ffx][repo] Waiting for background server to start
This adds waiting for the background server to start before returning
from `ffx repository server start --background`. This is usually not a
long time, but does make it more natural when writing test scripts
to start a background server without checking for the server to be
running.

Bug: 375636019
Change-Id: I12c15d9a5e34ffa5620a71536b55fa9ac655738a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1145136
Reviewed-by: Thomas Zander <thomasalpinus@google.com>
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>