[netstack3] Only rewrite port if necessary to avoid clash
In the case of implicit SNAT, where a connection is being NATed not
because the user has configured a NAT rule that applied to the first
packet of the flow, but only in order to prevent conflicts with other
tracked connections, only rewrite the source port if necessary to avoid
a tuple clash.

For connections that have NAT explicitly configured, such as through
Redirect or Masquerade NAT, we continue to ensure the reply tuple's port
falls in the specified range (if one is specified).

This makes implicit SNAT less invasive and also fixes an issue where UDP
packets with a source port of 0 are always rewritten, even when
unnecessary to avoid a conflict, because the default port ranges are
nonzero and thus never include the 0 port.

Test: netstack3-filter-test
      netstack3_udp_raw_syscall
      udp(-v6)_ns3_full_anvl_suite.sh -cases 2.3
Change-Id: I61b29b1759608cdd568ee4387655fda453e126f7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1145313
Fuchsia-Auto-Submit: Peter Johnston <peterjohnston@google.com>
Reviewed-by: Evan Wildenhain <wildenhain@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>