[fidl] Add type shape to external declarations
This unblocks Rust wire bindings, which need to know when external types
have out-of-line data. The type shape of a member would usually contain
enough information for this, but when external structs are optional the
field shape always contains out-of-line data (even when the struct
itself does not have out-of-line data).

Change-Id: I2c0628f11a346ead7abd20a51183389df6962c09
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1145253
Reviewed-by: Ian McKellar <ianloic@google.com>
Commit-Queue: David Koloski <dkoloski@google.com>