[kernel][arch] disable code coverage for early boot mmu code
For both arm64 and riscv64, disable generating code coverage on the
early boot mmu mapping code.

On arm64 in particular the compiler may generate unaligned stores to
the coverage array which will trigger an unaligned access since the
mmu is off and the data cache is disabled at this point.

For riscv64 it isn't an issue on current hardware, but a similar
condition may arise in the future so makes sense to disable it out
of an abundance of caution.

Change-Id: Idb6643bab86edd44a4dd7f0dd8862106ff944fbe
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1145432
Commit-Queue: Travis Geiselbrecht <travisg@google.com>
Reviewed-by: Nick Maniscalco <maniscalco@google.com>