[partition] Add Partition.GetMetadata
This returns all of the metadata for a partition that we'd expect to
find in a GPT.  For partition implementations that don't support some of
these fields, an absent value is returned.

This is redundant with the existing GetName/etc methods, but those have
been around forever and are too widely used to bother to port to this.

This will be used in the paver for preserving existing partitions across
a call to ResetPartitionTable.

Bug: 371060853
Change-Id: Id6809a9a557de4308d24add622435596509cccda
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1143900
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Reviewed-by: Chris Suter <csuter@google.com>
API-Review: Chris Suter <csuter@google.com>
Commit-Queue: James Sullivan <jfsulliv@google.com>