[tee] Defines TAs as `loadable_module()`s
This is rather than `shared_library()`'s, since TA's are meant to be
executables and not link inputs. This is of marginal build hygiene value
but also drops the perhaps-misleading prefix of "lib" from our compiled
TA filenames.

One consequence of doing s/shared_library/loadable_module is that we
lose the auto-generated .ifs link stubs created by our shared library
GN toolchain. This was recently realized to be of general diagnostic
value, as it gives insight into what subset of the TEE API is in used by
the TA. Accordingly, we begin to manually construct that in
fuchsia_ta_package() so as not to lose it. We also want this to
straightforwardly check at build-time that the official TA entrypoints
are indeed defined (to be done as a follow-on change).

Change-Id: Ia4a1a031a67b8fc9298a406161b5e8a2750ae34c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1143752
Reviewed-by: James Robinson <jamesr@google.com>
Commit-Queue: Joshua Seaton <joshuaseaton@google.com>
Fuchsia-Auto-Submit: Joshua Seaton <joshuaseaton@google.com>