[superproject] Roll antlion Update the import path of Honeydew WLAN types
Context: As part of fxr/1146114, we are changing the folder structure of
WLAN affordances in Honeydew to enable self servicable contributions. As
a result, as part of this CL, we are changing the old WLAN Honeydew
paths with the new ones.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/antlion/+/1142339
Original-Revision: 75ce3756f1ad91f533ddc9a2b6827e6e904770df
GitOrigin-RevId: 462209597c5cd7460b556133d82199be6fbf7704
Roller-URL: https://cr-buildbucket.appspot.com/build/8732780706660268929
CQ-Do-Not-Cancel-Tryjobs: true
Change-Id: I5397fdac6486672d02dd4335ce5b91885e79cb9b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1146211