[debugger] Add JobOnly filter config for FIDL clients
Adds the ability for FIDL clients to specify JobOnly attaching, so that
they may be notified of exceptions in any children of a matching
component, for example, without taking the exception channel of all
processes under the job.

Since DebugAgentServer now needs to keep track of the filters it
installs, we can also share some of the code with the zxdb frontend to
turn FilterMatch structs into correctly configured AttachRequests with
duplicate matches overlaying settings as we expect.

BUG: 362240229
BUG: 362240382
Change-Id: I9cdf5ebe969cb960552fde0b07e5c11dd8619390
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1143812
API-Review: Adam Barth <abarth@google.com>
Commit-Queue: Jacob Rutherford <jruthe@google.com>
Reviewed-by: Adam Barth <abarth@google.com>