[debugger] Fail loudly when symbol server auth fails
Previously, it was up to users to notice that their symbols were
missing, know to type `sym-server`, and also know how to interpret the
various states of the symbol servers.

With this change we remain silent in the good cases, but fail loudly
and give instructions on what to do in the common authentication failure
paths.

Change-Id: I1e1c96c0b0fa82f6bd69e2c72e305d49db96fe4b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1146874
Commit-Queue: Jacob Rutherford <jruthe@google.com>
Reviewed-by: Adam Barth <abarth@google.com>