[kernel][rppm] Clear pending power requests on domain changes.
Avoid sending a request intended for one power domain object to another
power domain when the request dispatch DPC races with changing the
power domain for the CPU.

The race condition is as follows:

1. A CPU is assigned domain D1.
2. A request R for a power level change becomes pending on the CPU,
   resulting in the power level DPC for the CPU being queued to dispatch
   the request.
3. Before the DPC acquires the scheduler queue lock for the CPU to latch
   R and D1, a different thread acquires the queue lock and replaces D1
   with D2 on the CPU.
4. The DPC acquires the queue lock and latches R and D2.

To prevent sending a request intended for the previous domain to the new
domain, we simply clear any pending request on the CPU when its domain
is changed. Since both the domain and request are protected by the queue
lock, the DPC cannot observe an inconsistent pairing.

Bug: b/375221894
Change-Id: I991a3b5b603abc2d59a9e3087c1f3818b29f6d1e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1146697
Reviewed-by: Gianfranco Valentino <gevalentino@google.com>
Fuchsia-Auto-Submit: Corey Tabaka <eieio@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>