[userboot] Fix multi-process/boot entry tests.
When userboot launches a process it will use the 'test' entry result to
determine sucess. This is fine 99% of the time, except when we have test
that involve being launches as a boot test, `userboot.next`.

When this happens, if there is a stub (multi process test) or there is
just the boot program performing a test, the result will either always
be sucess (default int value) or the stubs return value (0).

Change-Id: Ib4fded4d773087d445ffb23c8fc055961b1f939f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1143214
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Gianfranco Valentino <gevalentino@google.com>