[test] Prepare to support stable names for freeform metrics files
Return JSON | (str, JSON) from `process_freeform_metrics()` to allow callers in other repos to
tolerate both new and legacy behavior of `process_freeform_metrics()`

It turns out that the name of the file in which freeform metrics are
stored on test hosts is actually exposed in the database where the
metrics are stored. As such, dashboards might rely upon it. Using the
classname of the metrics processor, therefore, could cause outages in
the event of a refactor.

The most foolproof way to ensure that the author of a metrics processor
thinks about this issue and provides a stable identifier is to make the
filename a part of the return value from `process_freeform_metrics()`.

Bug: b/376097015
Test: CQ
Change-Id: I2901b6eae6a158f7b01431cc9918d73a74edd105
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1146197
Reviewed-by: Mark Seaborn <mseaborn@google.com>
Commit-Queue: Chris Masone <cmasone@google.com>