[netlink] Install routes in non-main tables
When a route needs to be installed in a non-main table, the netlink
worker now provisions a new fuchsia.net.routes.admin.RouteTable to hold
it.

Because rules aren't properly supported yet (see b/358649849), we still
need to write the routes to the main table as well.

This adds some complication. To cope with this, refactor things so that
the routes are held at rest as _FIDL_ route types rather than Netlink
ones. Because we're now installing routes in separate tables, the FIDL
Route structs that we observe through the watchers will now carry info
about which tables they're installed in, which aids us in making this
simplification.

This makes it easier to consume the FIDL routes watcher, as we can
always apply the watcher's events to our stored view of installed FIDL
routes, and then derive the netlink view of routes from that.

Bug: 358649849
Change-Id: I8b20d91d6815891f7d5165e3c7899b6b89272803
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1132574
Commit-Queue: Evan Wildenhain <wildenhain@google.com>
Reviewed-by: Jeff Martin <martinjeffrey@google.com>