[bazel_sdk] Override host tool paths when invoking ffx (again, again).
This is a reland of [1], which always includes the SDK manifest as
a runtime input, to avoid breaking OOT builds like the Android one.

Ensure that `ffx` and its plugins do not rely on the SDK manifest
to locate the host tool they may invoke at runtime, by passing their
location directly on the command line.

New functions are added to `ffx_tool.bzl` to retrieve the start
of the command-line invocation, that includes the necessary
`--config "sdk.override.<tool1>=<path1>,sdk.override.<tool2>=<path2>"`
argument.

This is required to allow moving the generation of the IDK to the
Bazel workspace. In this scenario, host tools will be built on-demand
as Bazel targets, and their location cannot be predicted before the
build, making it difficult to write them to the SDK manifest in advance.

(The SDK manifest will instead contain Bazel labels such as
`@fuchsia_idk//tools/x64:cmc` that `ffx` definitely cannot convert to
a proper host path on its own).

As two previous attempts at this CL had to be reverted, this has been
tested manually with an OOT Android Starnix-based build.

[1]: https://fuchsia-review.googlesource.com/c/fuchsia/+/1115539

Bug: 367274537
Bug: 287355615
Change-Id: I12dc7c33660ff963863a098e84131bd5d0075605
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1123613
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: David Turner <digit@google.com>
Reviewed-by: David Iserovich <diserovich@google.com>