[syslog] Remove syslog_internal sdk_component_manifests
This target, which was added in https://fxrev.dev/545048 and uses the
"internal" SDK category, is not in the SDK and all dependencies in that
CL have been removed. https://fxrev.dev/1121570 added a new target that
is in the "partner" IDK.

Remove the only two dependencies on this target. These dependencies are
on the directory, and this target happens to be the default target.

Bug: 345827642, 372986936
Change-Id: Ia80666c720226f01c7641dbbe5c21ffe0c4dc8ff
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1144834
Commit-Queue: David Dorwin <ddorwin@google.com>
Reviewed-by: Miguel Flores <miguelfrde@google.com>
Fuchsia-Auto-Submit: David Dorwin <ddorwin@google.com>
Reviewed-by: Brian Bosak <bbosak@google.com>
Reviewed-by: Matt Thiffault <mthiffault@google.com>