[shell] Route shell commands via shell-command packages
Rather than route /bin via the pkg-cache, it is now routed via the
pkg-shell-commands package's contents. Additionally, a new directory
boot-bin was added with the contents of boot-shell-commands. As
tools from /boot/bin are packaged, they will find themselves in that
directory instead.

This requires adding a the /boot-bin path to the global PATH so that
it is looked at for executables.

Change-Id: I0afc81363476a14a2846c81cfbbd2ce37ef7a7f1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1144037
Reviewed-by: Ben Keller <galbanum@google.com>
Fuchsia-Auto-Submit: Suraj Malhotra <surajmalhotra@google.com>
Reviewed-by: Aaron Wood <aaronwood@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Adam Barth <abarth@google.com>