[fake-pdev] Use rvalue ref for SetConfig
This will help ensure better errors are output when the user accidently
supplies and lvalue instead. Normally we would just delete the copy
constructor and implement the move constructor, but that's not helpful
here as we lose aggregate initialization for the config in that case.

Change-Id: Ibaf8490401ea098ecc29b02427b1a1e030444099
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1145961
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Sam Hansen <hansens@google.com>
Fuchsia-Auto-Submit: Suraj Malhotra <surajmalhotra@google.com>