[kernel][rppm] Fix `zx_system_set_processor_power_state`.
The syscall itself is fine, but this expression here is wrong.
The bucket is INCLUSIVE, but the expression is exclusive. So we get to
the point where we need to set the power level and do nothing.

E.g.
With 10 CPUs for example, and a bucket size of 64, we get :
10/64 = 0. That is bucket 0 contains all CPUs, but the expression above
would prevent from ever visiting this bucket.

Bug: 347701043
Change-Id: Id5d13e69e08ce2b691c0b8548e05cba4dc9c3a43
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1146753
Commit-Queue: Gianfranco Valentino <gevalentino@google.com>
Reviewed-by: Corey Tabaka <eieio@google.com>