[netstack3] Refactor NAT verdict types
Currently ConfigureNatResult is a struct that holds both a verdict
(typically either Accept or Drop) and a boolean `should_nat` value.
However, the latter is only meaningful in the context of an Accept
verdict; if the packet is to be dropped, its NAT configuration doesn't
matter. Invert the relationship between these two types such that
Verdict::Accept holds a value, which, in the case of NAT configuration,
is a NatConfigurationResult enum that indicates whether NAT should be
performed.

Change-Id: I51b4172f0a9b68cee6ac1ca45958a287cbca9fea
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1145959
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Peter Johnston <peterjohnston@google.com>
Reviewed-by: Peter DeLong <pdelong@google.com>