[test] Support stable names for freeform metrics files
Return filename and metrics directly from `process_freeform_metrics()`

It turns out that the name of the file in which freeform metrics are
stored on test hosts is actually exposed in the database where the
metrics are stored. As such, dashboards might rely upon it. Using the
classname of the metrics processor, therefore, could cause outages in
the event of a refactor.

The most foolproof way to ensure that the author of a metrics processor
thinks about this issue and provides a stable identifier is to make the
filename a part of the return value from `process_freeform_metrics()`.

Bug: b/376097015
Test: CQ
Change-Id: Idd48d3d35f5bae4836ae0580999a746497bf9e4c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1146979
Reviewed-by: Mark Seaborn <mseaborn@google.com>
Reviewed-by: Pierre Labatut <plabatut@google.com>
Commit-Queue: Chris Masone <cmasone@google.com>