[starnix][manager] Log which proxy wakes us
Log which of the proxied channels are waking us back up.
This makes it much easier to debug why the kernel is waking up.

Update the EventPairs to be in a HashSet so we can remove them
in O(1) without checking every Koid.

Change-Id: I36b38f537d35619524c0ca015a3ebf79924f4a39
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1146312
Fuchsia-Auto-Submit: David Gilhooley <dgilhooley@google.com>
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>