[netstack3] Skip implicit SNAT if conflict is identical
Currently if two identical packets end up creating exclusive connection
tracking entries and then race to finalize the entries in the conntrack
table, the second packet will see the first packet's connection in the
table and attempt to avoid a conflict by performing implicit SNAT. This
is not really what we want if the connection in the table and the
connection to be inserted are actually identical for all intents and
purposes; in this scenario, we want the second packet to simply adopt
the connection that's already inserted in the table.

Change the implicit SNAT logic to do this.

Fixed: 375012282
Test: netstack3-filter-test
Multiply: netstack-socket-integration-test
Change-Id: I68aab1c0f4bddcfe4decb0c3d7f74de6714f5c40
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1145960
Reviewed-by: Peter DeLong <pdelong@google.com>
Commit-Queue: Peter Johnston <peterjohnston@google.com>
Fuchsia-Auto-Submit: Peter Johnston <peterjohnston@google.com>