[build][toolchain] enable linux_arm
In order to build tests and helper binaries for 32-bit restricted
mode and starnix testing, this change maps in support for the
linux_arm toolchain (armv7-a hard float).

It is supports vdso32 and zircon integration test building, but core llvm testing is missing.

Bug: 372499348
Change-Id: Ic9ab5556b068ed83c48706fa47efe3345bfcdb78
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1124112
Reviewed-by: Petr Hosek <phosek@google.com>
Reviewed-by: Anirudh Mathukumilli <rudymathu@google.com>
Commit-Queue: Will Drewry <drewry@google.com>