[wlan] Use Policy-provided BSS desc. during roam.
In a WLAN Policy-initiated roam attempt, the BSS info of the target BSS
is provided directly to Fullmac. This is similar to the way that the
regular Connect works. This CL changes the logic in brcmfmac to use the
Policy-provided target BSS info rather than asking the firmware for it.

Note that this CL removes a test case from the brcmfmac_sim_test
reassoc_test, because that failure case no longer applies once brcmfmac
is using the Policy-provided BSS info. This test also passes using
Antlion suite WlanPolicyInitiatedRoamTest.

Test: fx test brcmfmac_sim_test

Change-Id: I0073591cc0a56dc03d085331db50daa737eefc3c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1146574
Reviewed-by: Karthik Krishnaswami <karthikrish@google.com>
Commit-Queue: Karl Ward <karlward@google.com>