[ffx][config] Adding $SHARED_DATA placeholder
This adds $SHARED_DATA to the list of strings that are interpolated
when resolving configuration values. The $SHARED_DATA is used for
persisting data that is shared between isolates and ffx instances on
the current host machine for the current user.

This is intended to be used for data that is used to manage global
resources such as TCP/IP port usage.

Bug: 374085877
Change-Id: I7572fb4cf59b737675e0c48fa3c079663e025edf
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1140397
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>
Reviewed-by: Steven Grady <slgrady@google.com>