[fidl][rust] Defer to IR on whether type is static
The IR contains information about whether a type can contain any
out-of-line data, so we can use that to replace the manual type
staticness calculation we were doing. This does add some complexity, but
in some ways it's good complexity. Wire unions with members which are
all inline-able can now be cloned, and unnecessary lifetimes on unions
and accessors are now removed.

Change-Id: I58d481eb9b0951ae2032235c9c336bf489fc34ca
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1143417
Reviewed-by: Ian McKellar <ianloic@google.com>
Commit-Queue: David Koloski <dkoloski@google.com>