[ffx][injector] Make injector optional
This changes the type of the FhoEnvironment::injector to
be Option<Arc<dyn Injector>>. This is in preparation for having subtools
that do not interact with the daemon, or even do not interact with any
Fuchsia device.

The next step would be to make an execution path that creates a
FhoEnvironment with the injector == None.

Change-Id: I8e7fcfc4609a2f2f18948ac6a32d75d656e94ace
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1147592
Reviewed-by: Steven Grady <slgrady@google.com>
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>