[logs] Route LogSink to gtest tests
Route LogSink to gtest tests.
This isn't normal routing, and just forwards
the entire /svc directory, which is a temporary
workaround for logs not working in this context.
The workaround has the additional consequence of
logs being misattributed to the runner instead
of the component that is running.

Bug: b/376735013
Change-Id: Ia83b927ff4704a11c5498fb15d60b62eedc309e1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1148816
Commit-Queue: Brian Bosak <bbosak@google.com>
Reviewed-by: Ankur Mittal <anmittal@google.com>