[utest] Fix flake in Timer.Restart
The Timer.Restart core test was inherently racy and could fail if the
test thread was preempted for longer than the timer deadline in between
when the timer was set and when the wait_one was invoked. This CL fixes
the issue by setting the timer deadline to a year in the future; while a
race technicallys till exists, the test thread should never be preempted
for that long.

Multiply: core-tests-ticks_get_force_syscall
Fixed: 375681700
Change-Id: I55646e38aec74ac3b200108b0cc6711250f8631a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1148133
Fuchsia-Auto-Submit: Anirudh Mathukumilli <rudymathu@google.com>
Reviewed-by: John Grossman <johngro@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>