[kernel][vm] Destructor helper for lock annotations
Moves most of the VmObjectPaged destructor into a helper method so that
it checks lock annotations. This is useful because normally a
destructor is considered not to need analysis as it runs when it is
just being cleaned up. However, in this case the VmObjectPaged can
still be found from different lists and so the destructor must still
take part in correct locking behavior.

As part of moving to a helper method some missing lock assertions are
added.

Change-Id: I32a1d0f68f88e1764bfcca3a5206f0a64671b29c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1135192
Reviewed-by: Sage Barreda <sagebarreda@google.com>
Commit-Queue: Adrian Danis <adanis@google.com>