[trace-provider] Silence warning about registration
On builds where trace_manager is not included, we're expected to fail to
register with trace manager. Let's not emit a warning in trace provider
so that we don't spam the log with every single trace provider on the
system warning it can't connect.

Bug: 376703561

Change-Id: I81d67679555c30d772c6ce3ba67481df39c009c1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1148234
Fuchsia-Auto-Submit: Gwen Mittertreiner <gmtr@google.com>
Reviewed-by: Krithika Balan <krithikabalan@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>