[kernel][vm] Remove unused ZeroPagesLocked optimization
This removes a specific optimization of the zero operation that made
zeroing the tail end of the clone of an anonymous VMO slightly more
efficient. The benefit is small, present users do not manage to hit
this case, and the usage of ReleaseCowParentPagesLocked here is
different compared to other use cases and complicates future hierarchy
changes.

Bug: 42058561
Change-Id: Ifd5cacf6eb877dc42e5f9382297c60b815d19e2c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1149013
Reviewed-by: Rasha Eqbal <rashaeqbal@google.com>
Fuchsia-Auto-Submit: Adrian Danis <adanis@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>