[sestarnix] Introduce AccessDecision to query APIs
AccessDecision encapsulates the allowed, audit-allow and audit-deny
bits for each permission, and a flags field which is currently used
to indicate whether or not the source domain is listed in the
policy's permissive domains map.

SecurityServer permission-check tests are updated to validate that
the expected audit decisions are made, and test cases are added
for the handling of enforcing, domain-permissive and global-
permissive access checks against kernel classes/permissions that
are not defined in the policy.

Bug: 362706116,
Change-Id: I7310b50cac1ecf922c38250ce6e71c235b19f5bd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1147712
Reviewed-by: Mark Dittmer <markdittmer@google.com>
Fuchsia-Auto-Submit: Wez <wez@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>